revert: add back in app.vue in /package/bootstrap*
The playground environment is great for simulating a user environment, however, as I have stated, it doesn't really work that well for developing components since playground only uses ./dist of the main package. It would need to be rebuilt constantly in order to see a good affect. There could be a workaround that I don't know about. Read the TODO in app.vue and make a discussion (or post on PR) if you think app.vue development area should be reverted or expanded like it was pre-pnpm/workspace commit