refactor(router.ts): change name and strongly typ
The name change was because 'router.ts' is confusing when an app uses createRouter, such as in the playground. It only contains one function, so it was made to export default.

chore: add note about bbutton

BButton has many assignments that are not computed values. I'm not sure why this is. It seems that if someone intentionally v-bind to an href on BButton, starts undefined, then puts on a link, that it would not react. Nor does the isLink in BButton use the isLink.ts util.

chore: update pnpm-lock

fix(BBadge): text-decoration-none when link.value

feat: begin adding booleanish option

This is to assist in user choice. At this time, it only applies to simple boolean types, not complex i.e boolean | any. It may go through some changes. A user would be able to do prop="true" vs requiring :prop="true". At this time, I'm unsure if there's a better, less boilerplatish method of converting the Booleanish prop to boolean.