chore: update test utils
I would like to migrate to happy-dom instead of jsdom, as it seems to be near 50% faster, which, when all components start containing many tests, could be a substantial time saver. However, I haven't been able to get it to work out of the box. So I will probably try later.