Update package.json
This simply fixes the doc dev due to the mismatch in dependencies. Please note going forward that we should start to plan less frequent releases. This gives enough time to fix any trailing issues and review code. 

Moving forward changes to dependencies and other aspects of the package.json or config files should not be modified and pushed so lightly without extensively reviewing their impact on other parts of the codebase.