feat: finish bnavbartoggle and navitem
BNavbarToggle uses the v-b-toggle directive, it doesn't seem that this directive contains any built in functionality for controlling the disabled state. I am unsure of how to make the directive not expand or contract on disabled state. Since there is only one other example usage of v-b-toggle in the lib, I have made a hacky workaround that when disabled, the targetId is just undefined.

refactor: move bnavbar content to separate folder