fix: include explicit imports of components
This caused an issue where in tests, and the possibility of single imports of specific components, (rather than relying on global imports) would cause a potential error. For example, string 'BLink' in a dynamic <component :is>, would not be resolved in mounted tests, or in a single import of that component. This causes nearly invisible errors in tests, unless using the wrapper.getComponent function to test to ensure it exists