test: more components read TODOs
Some components are not done. Read their TODOS. Some components used heavy reliance on .find()... .find() should be reserved for proving the existance of a component. Preference is to use .get() if the component exists, as it will throw an error, according to vue-test-utils.

fix: BCol classes persisting after prop change

test: refactor .find to use .get