test: remove waitNT util in favor of direct vue
When using TS, the waitNT function would break because they aren't the same types. I intend for all the tests to be converted to TS by the end of testing so it would be best if tests didn't break the lib.

test: BToast tests were decomposed to have multiple, more focused tests

test: toaster

test: transition not done only a template file

fix: BToaster direct import of BToast

feat: BCloseButton dynamic ariaLabel prop